requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
psycopg2-binary==2.9.9
redis==5.0.4
pydantic==2.7.0
sqlalchemy==2.0.30
sentence-transformers==2.7.0
prometheus-client==0.20.0
tenacity==9.0.0
pytest==8.2.0
requests==2.32.3
